Displays a list of custom post type "video" in a portfolio style.
jQuery is used to slide in a "play" and "more info" button over the video posts featured image. When the user clicks on play or info it will open the video or show the description in a fancy box pop up.
This site uses css media queries and "fittext" jquery plugin for responsive headlines, to accomplish a responsive design.

See the Wordpress jQuery file for corresponding code.